remove population of meta fields from tab-filter, to put it in grid service
