+++ projects/safe/src/lib/components/query-builder/tab-filter/tab-filter.component.ts
@@ -2,6 +2,7 @@ import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
 import { FormArray, FormBuilder, FormGroup } from '@angular/forms';
 import { clone, isDate } from 'lodash';
 import { SafeApiProxyService } from '../../../services/api-proxy.service';
+import { SafeGridService } from '../../../services/grid.service';
 import { QueryBuilderService } from '../../../services/query-builder.service';
 
 /**
@@ -195,12 +196,12 @@ export class SafeTabFilterComponent implements OnInit {
    *
    * @param formBuilder This is the service that will be used to build forms.
    * @param queryBuilder This is the service that will be used to build the query.
-   * @param apiProxyService This is the service that will be used to make the API call.
+   * @param gridService Shared grid service
    */
   constructor(
     private formBuilder: FormBuilder,
     private queryBuilder: QueryBuilderService,
-    private apiProxyService: SafeApiProxyService
+    private gridService: SafeGridService
   ) {}
 
   ngOnInit(): void {
@@ -217,13 +218,13 @@ export class SafeTabFilterComponent implements OnInit {
           for (const field in res.data) {
             if (Object.prototype.hasOwnProperty.call(res.data, field)) {
               this.metaFields = Object.assign({}, res.data[field]);
-              await this.populateMetaFields();
+              await this.gridService.populateMetaFields(this.metaFields);
             }
           }
         });
       }
     } else {
-      this.populateMetaFields();
+      this.gridService.populateMetaFields(this.metaFields);
     }
     this.form.value?.filters.forEach((x: any, index: number) => {
       if (x.field) {
@@ -248,71 +249,6 @@ export class SafeTabFilterComponent implements OnInit {
         this.selectedFields.splice(index, 1, {});
       }
     });
-  }
-
-  /**
-   * Fetch choices from URL if needed
-   */
-  private async populateMetaFields(): Promise<void> {
-    const promises: Promise<any>[] = [];
-    for (const fieldName of Object.keys(this.metaFields)) {
-      const meta = this.metaFields[fieldName];
-      if (meta.choicesByUrl) {
-        const url: string = meta.choicesByUrl.url;
-        const localRes = localStorage.getItem(url);
-        if (localRes) {
-          this.metaFields[fieldName] = {
-            ...meta,
-            choices: this.extractChoices(
-              JSON.parse(localRes),
-              meta.choicesByUrl
-            ),
-            choicesByUrl: null,
-          };
-        } else {
-          promises.push(
-            this.apiProxyService
-              .promisedRequestWithHeaders(url)
-              .then((value: any) => {
-                localStorage.setItem(url, JSON.stringify(value));
-                this.metaFields[fieldName] = {
-                  ...meta,
-                  choices: this.extractChoices(value, meta.choicesByUrl),
-                  choicesByUrl: null,
-                };
-              })
-          );
-        }
-      }
-    }
-    await Promise.all(promises);
-  }
-
-  /**
-   * Extracts choices using choicesByUrl properties
-   *
-   * @param res Result of http request.
-   * @param choicesByUrl Choices By Url property.
-   * @param choicesByUrl.path Path of the choice
-   * @param choicesByUrl.value Value of the choice
-   * @param choicesByUrl.text Text of the choice
-   * @returns list of choices.
-   */
-  private extractChoices(
-    res: any,
-    choicesByUrl: { path?: string; value?: string; text?: string }
-  ): { value: string; text: string }[] {
-    const choices = choicesByUrl.path ? [...res[choicesByUrl.path]] : [...res];
-    return choices
-      ? choices.map((x: any) => ({
-          value: (choicesByUrl.value ? x[choicesByUrl.value] : x).toString(),
-          text: choicesByUrl.text
-            ? x[choicesByUrl.text]
-            : choicesByUrl.value
-            ? x[choicesByUrl.value]
-            : x,
-        }))
-      : [];
   }
 
   /**
