+++ projects/safe/src/lib/services/grid.service.ts
@@ -237,6 +237,7 @@ export class SafeGridService {
    * @param metaFields List of meta fields
    */
   public async populateMetaFields(metaFields: any): Promise<void> {
+    const promises: Promise<any>[] = [];
     for (const fieldName of Object.keys(metaFields)) {
       const meta = metaFields[fieldName];
       if (meta.choicesByUrl) {
@@ -249,18 +250,25 @@ export class SafeGridService {
               JSON.parse(localRes),
               meta.choicesByUrl
             ),
+            // choicesByUrl: null,
           };
         } else {
-          const res: any =
-            await this.apiProxyService.promisedRequestWithHeaders(url);
-          localStorage.setItem(url, JSON.stringify(res));
-          metaFields[fieldName] = {
-            ...meta,
-            choices: this.extractChoices(res, meta.choicesByUrl),
-          };
+          promises.push(
+            this.apiProxyService
+              .promisedRequestWithHeaders(url)
+              .then((value: any) => {
+                localStorage.setItem(url, JSON.stringify(value));
+                metaFields[fieldName] = {
+                  ...meta,
+                  choices: this.extractChoices(value, meta.choicesByUrl),
+                  // choicesByUrl: null,
+                };
+              })
+          );
         }
       }
     }
+    await Promise.all(promises);
   }
 
   /**
@@ -302,6 +310,7 @@ export class SafeGridService {
         text: choicesByUrl.otherText ? choicesByUrl.otherText : 'Other',
       });
     }
+    choices.sort((a: any, b: any) => a.text.localeCompare(b.text));
     return choices;
   }
